+++ packages/foundryup/src/extract.ts
@@ -143,6 +143,9 @@ async function extractFromTar(
     extractTar(
       {
         cwd: dir,
+        // @ts-expect-error: Types here broke in `tar@7.5.12`, but it appears to
+        // work fine as-is.
+        // See: https://github.com/isaacs/node-tar/issues/459
         transform: (entry) => {
           const absolutePath = entry.absolute;
           if (!absolutePath) {